+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.2.0
 * Add option to disable type checking on isObjectLike  
 * Add isPopulatedObject  
+* Add isDate, isDateGreaterThan, and isDateLessThan  
 
 # v1.1.1  
 * Add debug option to isObjectLike